+++ src/main/java/com/lidiabazhenova/webapp/controllers/product/EditProductsServlet.java
@@ -43,7 +43,7 @@ public class EditProductsServlet extends HttpServlet {
                 .setProductUrl(request.getParameter("productUrl"))
                 .setProductName(request.getParameter("productName"))
                 .setProductPrice(Double.parseDouble(request.getParameter("productPrice")))
-                .setProductQuantity(Double.parseDouble(request.getParameter("productQuantity")))
+                .setProductQuantity(Integer.parseInt(request.getParameter("productQuantity")))
                 .build();
 
         try {
